flask_voicemagic/requirements.txt
flask>=0.12.2
pillow>=4.1.1